Commit message: Add Dockerfile for Python project
A Dockerfile was created for the Python project to automatize the building of the Docker image. It uses Python 3.10.12 as a base image, sets the working directory, copies over the source code and installs necessary dependencies from requirements.txt. This will help in having a consistent environment for reproducible results.